+++ .github/genbot/__main__.py
@@ -1,4 +1,4 @@
-"""Entry point for `python -m genbot`."""
+"""Entry point for `python -m genbot`"""
 
 from genbot.cli import main
