+++ .github/genbot/template.py
@@ -1,10 +1,10 @@
-"""Template rendering for genbot."""
+"""Template rendering for genbot"""
 
 from pathlib import Path
 
 
-def render(template_path: Path, robot: str, **extras) -> str:
-    """Read a template and substitute __ROBOT__ (and any extras) into it"""
+def populate_template(template_path: Path, robot: str, **extras) -> str:
+    """Read a template and substitutes __ROBOT__"""
     text = template_path.read_text()
     text = text.replace("__ROBOT__", robot)
     for key, value in extras.items():
@@ -15,4 +15,4 @@ def render(template_path: Path, robot: str, **extras) -> str:
 def write_template(src: Path, dest: Path, robot: str, **extras) -> None:
     """Writes a template file"""
     dest.parent.mkdir(parents=True, exist_ok=True)
-    dest.write_text(render(src, robot, **extras))
+    dest.write_text(populate_template(src, robot, **extras))
