+++ .vscode/settings.json
@@ -77,11 +77,13 @@
     "python.analysis.extraPaths": [
         "/opt/ros/humble/lib/python3.10/site-packages",
         "/opt/ros/humble/local/lib/python3.10/dist-packages",
-        "robot-sim/sim_common"
+        "robot-sim/sim_common",
+        ".github"
     ],
     "python.autoComplete.extraPaths": [
         "/opt/ros/humble/lib/python3.10/site-packages",
         "/opt/ros/humble/local/lib/python3.10/dist-packages",
-        "robot-sim/sim_common"
+        "robot-sim/sim_common",
+        ".github"
     ],
 }
